Complete JWT authorized and CRUD Functionality

+++ backend/controllers/goalController.js
@@ -1,19 +1,28 @@
 import { StatusCodes } from "http-status-codes";
 import Goal from "../models/goalModel.js";
+import User from "../models/userModel.js";
 import { createCustomAPIError } from "../errors/custom-error.js";
 
 //Desc Get All Goals
 //Route /api/v1/goals (GET)
 //Access Private
 const getGoals = async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user.id });
   res.status(StatusCodes.OK).json({ nbHits: goals.length, goals });
 };
-//Desc Post Goals
+//Desc Set Goals
 //Route /api/v1/goals (POST)
 //Access Private
-const setGoals = async (req, res) => {
-  const setGoal = await Goal.create(req.body);
+const setGoals = async (req, res, next) => {
+  if (!req.body.text) {
+    return next(
+      createCustomAPIError("Please Add Text Field", StatusCodes.BAD_REQUEST)
+    );
+  }
+  const setGoal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  });
   res.status(StatusCodes.CREATED).json({ setGoal });
 };
 //Desc Get Single Goals
@@ -35,6 +44,23 @@ const getSingleGoals = async (req, res, next) => {
 const updateGoals = async (req, res, next) => {
   const { id } = req.params;
   const data = req.body;
+  const goal = await Goal.findById(id);
+  if (!goal) {
+    return next(createCustomAPIError("Goal Not Found", StatusCodes.NOT_FOUND));
+  }
+  const user = await User.findById(req.user.id);
+  //Check for login user
+  if (!user) {
+    return next(
+      createCustomAPIError("User Not Found", StatusCodes.UNAUTHORIZED)
+    );
+  }
+  //Make Sure the Logged in user Matches the Goal User
+  if (goal.user.toString() !== user.id) {
+    return next(
+      createCustomAPIError("User Not Authorized", StatusCodes.UNAUTHORIZED)
+    );
+  }
   const updatedGoal = await Goal.findByIdAndUpdate(id, data, {
     new: true,
     runValidators: true,
@@ -51,6 +77,25 @@ const updateGoals = async (req, res, next) => {
 //Access Private
 const deleteGoals = async (req, res, next) => {
   const { id } = req.params;
+  const goal = await Goal.findById(id);
+  if (!goal) {
+    return next(createCustomAPIError("Goal Not Found", StatusCodes.NOT_FOUND));
+  }
+
+  const user = await User.findById(req.user.id);
+  //Check for login user
+  if (!user) {
+    return next(
+      createCustomAPIError("User Not Found", StatusCodes.UNAUTHORIZED)
+    );
+  }
+  //Make Sure the Logged in user Matches the Goal User
+  if (goal.user.toString() !== user.id) {
+    return next(
+      createCustomAPIError("User Not Authorized", StatusCodes.UNAUTHORIZED)
+    );
+  }
+
   const deletedGoal = await Goal.findByIdAndDelete(id);
   if (!deletedGoal) {
     return next(
